Enable infinite regions by default
NOTE: There are some who have indicated that this creates issues with viewers.  After testing this I find no issues created for the viewer.  All this does is just allows the user to go outside the region even if there is not a neighboring region.  This has no affects on on any region management or parcel management or anything else.

Signed-off-by: EmperorStarfinder <emperor@secondgalaxy.com>