Fix Version Information
Note: We are aware of warnings being thrown on the grid and region servers.  This is due to a mistake made in the last commit.  There is no need to report any bugs on this as we are aware of the issue and are in the process of correcting this.  The commit with the corrections will be pushed to master shortly.

Signed-off-by: EmperorStarfinder <emperor@secondgalaxy.com>