Clean up the UserAccountService
- Remove the reset rezday command
- Make sure that rez dates cant be altered for accounts imported from .csv viles if no rezdate exists in the file
- General Cleanup
- A couple of quick vises in voice or build would break

Note: You should never alter the rezdate of a user.  This is bad practice and might also have unintended impacts on the user with other grid and region services.  Similarly the system will create a rezdate for the user automatically.  If you export the useraccounts to a .csv file for later import be sure you include the rezdate or you might cause unintended problems for that user if you don't.

Signed-off-by: BriakallistaStarfinder <briakallista@secondgalaxy.com>