A bunch of updates
- Fixes for rezzing objects
- Fixes for avatar rendering
- Fixes for OARs
- General code cleanup
- Merging in a number of updates (many thanks to everyone who worked on these)

NOTE: Linux traditionally will have a different number of warnings then Windows will due to how Linux handles the build process.  Most warnings are not related to zlib and will get fixed as we are able to fix them.  As always it is encouraged that if you use Virtual Universe in a production environment you should always use the current stable release.  For Linux this is 14.04.3 LTS.

Signed-off-by: EmperorStarfinder <emperor@secondgalaxy.com>