A little cleanup
NOTE: For those who run into an issue with the run configurator command on your Universe.exe when in grid mode not updating your database info in Config/Sim/Data/Data.ini and Config/Sim/Data/SQLite.ini or MySQL.ini please note that we are aware of this and will have a fix shortly.  In the meantime you will need to edit these files as stated to make your region servers work properly as the correct database is needed to connect your regions with their proper estates.

Signed-off-by: EmperorStarfinder <emperor@secondgalaxy.com>