Rebase Part 2: Fresh code
- Fixes Issue #49
- Cleans up the repository setting the code to Virtual Universe Version 1.0.2 RC2
- Ensures proper credit goes to sources where the code and files came from prior to arriving in Virtual Universe  Many thanks to all sources for the great work over the years
- Coverity problems fixed is now 245

NOTE: Users may need to start with a fresh database to ensure all correct tables are available to all users.

Note 2: Beginning with the development of Virtual Universe Version 1.0.3 the development level code will be in https://github.com/Virtual-Universe/Virtual-Dev in order to ensure users have a stable version of the code to work with while the next great version is being built.

Signed-Off-By: EmperorStarfinder<emperor@secondgalaxy.com>