Version 1.0.2 Release
This is the Release Version for Virtual Universe 1.0.2

Changes:

- Coverity fixes
- Code cleanup
- Code updates
- Updates from upstream providers

Important Note 1: Users and Grid Operators using Virtual Universe are strongly encourage to start with a fresh database and fresh copy of Virtual Universe to avoid false errors.

Important Note 2: When opening an issue please use the built in issue template provided.  By answering the questions you help us with the information we need to reproduce the bug or issue you are reporting.  This allows us to find and fix problems quickly.

Signed-Off-By: EmperorStarfinder<emperor@secondgalaxy.com>