Virtual Universe Version 1.0.3 Release Candidate 2
- Officially built versions for this release candidate will be available on http://virtual-planets.org shortly

NOTE: All newly opened feature requests will not be considered until after the official release of Virtual Universe Version 1.0.3 Release on June 19, 2017.  If the feature request was accepted during the course of development of Virtual Universe Version 1.0.3 Release then it will be included in the final release for this version or will be pushed to the Version 1.0.4 Milestone.

NOTE 2: From now until June 19, 2017 will be dedicated to bug fixes and tweaks prior to the official release of this version.

Signed-Off-By: EmperorStarfinder<emperor@secondgalaxy.com>